Add isServiceRunning().

+++ Source/src/com/phonegap/geolocationplugin/SqliteController.java
@@ -125,10 +125,10 @@ public class SqliteController extends SQLiteOpenHelper {
                     oneLoc.put("dt", dt);
                     oneLoc.put("lat", Double.valueOf(cursor.getString(2)));
                     oneLoc.put("lon", Double.valueOf(cursor.getString(3)));
-                    oneLoc.put("acc", Integer.valueOf(cursor.getString(4)));
-                    oneLoc.put("alt", Integer.valueOf(cursor.getString(5)));
-                    oneLoc.put("hdg", Integer.valueOf(cursor.getString(6)));
-                    oneLoc.put("spd", Integer.valueOf(cursor.getString(7)));
+                    oneLoc.put("acc", Double.valueOf(cursor.getString(4)));
+                    oneLoc.put("alt", Double.valueOf(cursor.getString(5)));
+                    oneLoc.put("hdg", Double.valueOf(cursor.getString(6)));
+                    oneLoc.put("spd", Double.valueOf(cursor.getString(7)));
                     locList.put(oneLoc);
                 } catch (JSONException e) {
                 	e.printStackTrace();
@@ -165,10 +165,10 @@ public class SqliteController extends SQLiteOpenHelper {
                     oneLoc.put("dt", dt);
                     oneLoc.put("lat", Double.valueOf(cursor.getString(2)));
                     oneLoc.put("lon", Double.valueOf(cursor.getString(3)));
-                    oneLoc.put("acc", Integer.valueOf(cursor.getString(4)));
-                    oneLoc.put("alt", Integer.valueOf(cursor.getString(5)));
-                    oneLoc.put("hdg", Integer.valueOf(cursor.getString(6)));
-                    oneLoc.put("spd", Integer.valueOf(cursor.getString(7)));
+                    oneLoc.put("acc", Double.valueOf(cursor.getString(4)));
+                    oneLoc.put("alt", Double.valueOf(cursor.getString(5)));
+                    oneLoc.put("hdg", Double.valueOf(cursor.getString(6)));
+                    oneLoc.put("spd", Double.valueOf(cursor.getString(7)));
                     
                     locList.put(oneLoc);
                 } catch (JSONException e) {
